Changes to typescript as main lang. Separate gcp functions to separate file.
